requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-bootstrap-pagination==1.7.1
django-bootstrap3==12.0.3
django-contrib-comments==1.9.2
django-dropbox-storage==0.4.1
django-import-export==2.0.2
django-pagination==1.0.7
django-social-share==1.3.2
django-storages==1.8
django-youtube==0.2
dropbox==9.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.42
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
urllib3==1.25.8
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
